+++ example/LeftLedAnimation.h
@@ -45,7 +45,7 @@ public:
         traceme;
     }
 
-    bool on(uint8_t requestorId) override
+    bool on(short requestorId) override
     {
         traceme;
 
@@ -66,7 +66,7 @@ public:
         return true;
     }
 
-    bool off(uint8_t requestorId) override
+    bool off(short requestorId) override
     {
         traceme;
