+++ example/LeftLedTurnConfig.h
@@ -18,7 +18,7 @@ public:
         traceme;
     }
 
-    bool on(uint8_t requestorId) override
+    bool on(short requestorId) override
     {
         traceme;
 
@@ -28,7 +28,7 @@ public:
         return true;
     }
 
-    bool off(uint8_t requestorId) override
+    bool off(short requestorId) override
     {
         traceme;
